LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-DOMWindowTimers-setTimeout-setInterval-expected.txt

CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

PASS window.setTimeout assigned via policy (successful Script transformation).
PASS `window.setTimeout(string)` throws.
PASS `window.setTimeout(null)` throws.
PASS window.setInterval assigned via policy (successful Script transformation).
PASS `window.setInterval(string)` throws.
PASS `window.setInterval(null)` throws.
PASS `setTimeout(string)`, `setInterval(string)` via default policy (successful Script transformation).
PASS `setTimeout(null)`, `setInterval(null)` via default policy (successful Script transformation).
